fix crash, fix statistics bug, new launchimages

+++ Classes/CSVParser.m
@@ -199,7 +199,7 @@ foundHeader:
         }
         else
         {
-            fieldName = [NSString stringWithFormat: @"FIELD_%ld", fieldCount + 1];
+            fieldName = [NSString stringWithFormat: @"FIELD_%ld", (long)(fieldCount + 1)];
             [fieldNames addObject: fieldName];
             fieldNamesCount++;
         }
